requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altgraph==0.17.4
attrs==25.1.0
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
keyboard==0.13.5
multidict==6.1.0
packaging==24.2
pefile==2023.2.7
propcache==0.2.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
setuptools==75.8.0
yarl==1.18.3

